refactor(server): split handlers.go into focused per-flow files
handlers.go had grown to 2110 lines mixing discovery, browser login, the token
endpoint and every grant. Split it into focused files (discovery, authorize,
login, approval, token, userinfo, and one grant_*.go per grant) with behaviour
preserved 1:1 — pure code movement, verified by the existing test suite.

Connector authorization now lives in connectorauth.go with a documented
invariant (client AllowedConnectors + connector GrantTypes must both be
enforced by every grant), so the checks are in one obvious place instead of
scattered inline. refreshhandlers.go is renamed grant_refresh.go for
consistency with the other grants.
