+++ server/authorize.go
@@ -1,0 +1,177 @@
+package server
+
+import (
+	"context"
+	"html/template"
+	"net/http"
+	"net/url"
+	"strings"
+
+	"github.com/dexidp/dex/storage"
+)
+
+// grantTypeFromAuthRequest determines the grant type from the authorization request parameters.
+func (s *Server) grantTypeFromAuthRequest(r *http.Request) string {
+	redirectURI := r.Form.Get("redirect_uri")
+	if redirectURI == deviceCallbackURI || strings.HasSuffix(redirectURI, deviceCallbackURI) {
+		return grantTypeDeviceCode
+	}
+	responseType := r.Form.Get("response_type")
+	for _, rt := range strings.Fields(responseType) {
+		if rt == "token" || rt == "id_token" {
+			return grantTypeImplicit
+		}
+	}
+	return grantTypeAuthorizationCode
+}
+
+// handleAuthorization handles the OAuth2 auth endpoint.
+func (s *Server) handleAuthorization(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	// Extract the arguments
+	if err := r.ParseForm(); err != nil {
+		s.logger.ErrorContext(r.Context(), "failed to parse arguments", "err", err)
+
+		s.renderError(r, w, http.StatusBadRequest, ErrMsgInvalidRequest)
+		return
+	}
+
+	connectorID := r.Form.Get("connector_id")
+	allConnectors, err := s.storage.ListConnectors(ctx)
+	if err != nil {
+		s.logger.ErrorContext(r.Context(), "failed to get list of connectors", "err", err)
+		s.renderError(r, w, http.StatusInternalServerError, "Failed to retrieve connector list.")
+		return
+	}
+
+	// Determine the grant type from the authorization request to filter connectors.
+	grantType := s.grantTypeFromAuthRequest(r)
+	connectors := make([]storage.Connector, 0, len(allConnectors))
+	for _, c := range allConnectors {
+		if GrantTypeAllowed(c.GrantTypes, grantType) {
+			connectors = append(connectors, c)
+		}
+	}
+
+	// Filter connectors based on the client's allowed connectors list.
+	// client_id is required per RFC 6749 §4.1.1.
+	client, authErr := s.getClientWithAuthError(ctx, r.Form.Get("client_id"))
+	if authErr != nil {
+		s.renderError(r, w, authErr.Status, authErr.Error())
+		return
+	}
+	connectors = filterConnectors(connectors, client.AllowedConnectors)
+
+	if len(connectors) == 0 {
+		s.renderError(r, w, http.StatusBadRequest, "No connectors available for this client.")
+		return
+	}
+
+	// We don't need connector_id any more
+	r.Form.Del("connector_id")
+
+	// Construct a URL with all of the arguments in its query
+	connURL := url.URL{
+		RawQuery: r.Form.Encode(),
+	}
+
+	// Redirect if a client chooses a specific connector_id
+	if connectorID != "" {
+		for _, c := range connectors {
+			if c.ID == connectorID {
+				connURL.Path = s.absPath("/auth", url.PathEscape(c.ID))
+				http.Redirect(w, r, connURL.String(), http.StatusFound)
+				return
+			}
+		}
+		s.renderError(r, w, http.StatusBadRequest, "Connector ID does not match a valid Connector")
+		return
+	}
+
+	if len(connectors) == 1 && !s.alwaysShowLogin {
+		connURL.Path = s.absPath("/auth", url.PathEscape(connectors[0].ID))
+		http.Redirect(w, r, connURL.String(), http.StatusFound)
+		return
+	}
+
+	// Skip connector selection if a valid session exists, unless prompt=select_account or alwaysShowLogin.
+	if s.sessionConfig != nil {
+		authReq, _, err := s.parseAuthorizationRequest(r)
+		if err != nil {
+			s.logger.ErrorContext(r.Context(), "failed to parse authorization request", "err", err)
+
+			switch authErr := err.(type) {
+			case *redirectedAuthErr:
+				authErr.Handler().ServeHTTP(w, r)
+			case *displayedAuthErr:
+				s.renderError(r, w, authErr.Status, err.Error())
+			default:
+				panic("unsupported error type")
+			}
+			return
+		}
+		prompt, err := ParsePrompt(authReq.Prompt)
+		if err != nil {
+			// Server error because authReq was validated before saving it to database.
+			s.redirectWithError(w, r, authReq, errServerError, "Invalid authentication request")
+			return
+		}
+
+		// Invalid prompts will be validated and properly redirected later
+		if !s.alwaysShowLogin && !prompt.SelectAccount() {
+			session := s.getValidSession(ctx, w, r)
+			if session != nil {
+				for _, c := range connectors {
+					if c.ID != session.ConnectorID {
+						continue
+					}
+					connURL.Path = s.absPath("/auth", url.PathEscape(session.ConnectorID))
+					http.Redirect(w, r, connURL.String(), http.StatusFound)
+					return
+				}
+			}
+		}
+		if prompt.None() {
+			// Cannot authenticate silently with prompt=none.
+			s.redirectWithError(w, r, authReq, errLoginRequired, "id_token_hint does not match authenticated user")
+			return
+		}
+	}
+
+	connectorInfos := make([]connectorInfo, 0, len(connectors))
+	for _, conn := range connectors {
+		connURL.Path = s.absPath("/auth", url.PathEscape(conn.ID))
+		connectorInfos = append(connectorInfos, connectorInfo{
+			ID:   conn.ID,
+			Name: conn.Name,
+			Type: conn.Type,
+			URL:  template.URL(connURL.String()),
+		})
+	}
+
+	if err := s.templates.login(r, w, connectorInfos); err != nil {
+		s.logger.ErrorContext(r.Context(), "server template error", "err", err)
+	}
+}
+
+// getClientWithAuthError retrieves a client by ID and returns a displayedAuthErr on failure.
+// Invalid client_id is not treated as a redirect error per RFC 6749 §4.1.2.1.
+// https://datatracker.ietf.org/doc/html/rfc6749#section-4.1.2.1
+func (s *Server) getClientWithAuthError(ctx context.Context, clientID string) (storage.Client, *displayedAuthErr) {
+	client, err := s.storage.GetClient(ctx, clientID)
+	if err != nil {
+		if err == storage.ErrNotFound {
+			s.logger.ErrorContext(ctx, "invalid client_id provided", "client_id", clientID)
+			return storage.Client{}, newDisplayedErr(http.StatusBadRequest, "Invalid client_id provided.")
+		}
+		s.logger.ErrorContext(ctx, "failed to get client", "client_id", clientID, "err", err)
+		return storage.Client{}, newDisplayedErr(http.StatusInternalServerError, "Database error.")
+	}
+	return client, nil
+}
+
+func (s *Server) renderError(r *http.Request, w http.ResponseWriter, status int, description string) {
+	if err := s.templates.err(r, w, status, description); err != nil {
+		s.logger.ErrorContext(r.Context(), "server template error", "err", err)
+	}
+}
